+++ src/components/fhrs/LocalAuthorityMap/localAuthorityMap.js
@@ -2,6 +2,7 @@ import './localAuthorityMap.html.twig';
 import './localAuthorityMap.scss';
 import placeholderMap from './assets/fhrs-map.svg';
 import { domContentLoadedWrapper } from '../../../helpers';
+import imageMapResize from 'image-map-resizer';
 
 export default function () {
   function callback() {
@@ -14,6 +15,9 @@ export default function () {
     // Set placeholder map
     const map = localAuthorityMap.querySelector('img');
     map.src = placeholderMap;
+
+    // Resize image map automatically
+    imageMapResize();
   }
   return domContentLoadedWrapper(callback);
 }